fixed false negative for floating point values

+++ demo/src/main/java/com/epam/demo/Utils.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class Utils {
     public static boolean isAllPositiveNumbers(List<String> args) {
         for (String string : args) {
-            if (!StringUtils.isNumeric(string) || Integer.parseInt(string) <= 0) {
+            if (!StringUtils.isNumeric(string) || Double.parseDouble(string) <= 0) {
                 return false;
             }
         }
